+++ src/components/inCreatePickleTwo/costSelect/categorySelect.tsx/CategorySelect.tsx
@@ -1,0 +1,5 @@
+export default function CategorySelect() {
+  return (
+    <div>CategorySelect</div>
+  )
+}
